requirements.txt
pandas
numpy
pyyaml
requests
scikit-learn
sentence-transformers
rouge-score
matplotlib
seaborn
scipy
statsmodels
